chore: run identify-codeowners with node, drop tsx dep
