+++ .github/scripts/shared/pull-request.mts
@@ -67,41 +67,3 @@ export async function retrievePullRequest(
 
   return pullRequest;
 }
-
-export type PullRequestFile = {
-  filename: string;
-  additions: number;
-  deletions: number;
-}
-
-/**
- * Retrieves files changed in a specific pull request
- * @param octokit GitHub API client
- * @param repoOwner Repository owner (e.g., "MetaMask")
- * @param repoName Repository name (e.g., "metamask-extension")
- * @param prNumber Pull request number
- * @returns Array of filenames that were changed in the PR
- */
-export async function retrievePullRequestFiles(
-  octokit: InstanceType<typeof GitHub>,
-  repoOwner: string,
-  repoName: string,
-  prNumber: number,
-): Promise<PullRequestFile[]> {
-  // `octokit.rest.pulls.listFiles` paginates at 30 files per page by default.
-  // Use `octokit.paginate` so PRs with more than one page of files (e.g.
-  // large refactors) return every changed file, otherwise downstream
-  // codeowner matching silently sees only the first page.
-  const files = await octokit.paginate(octokit.rest.pulls.listFiles, {
-    owner: repoOwner,
-    repo: repoName,
-    pull_number: prNumber,
-    per_page: 100,
-  });
-
-  return files.map((file) => ({
-    filename: file.filename,
-    additions: file.additions || 0,
-    deletions: file.deletions || 0,
-  }));
-}
